requirements/req-dev.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements/req-dev.txt requirements/req-dev.in
#
annotated-types==0.5.0
    # via pydantic
asgiref==3.7.2
    # via django
black==23.1.0
    # via -r requirements/req-dev.in
build==0.10.0
    # via pip-tools
click==8.1.5
    # via
    #   black
    #   pip-tools
django==4.2.3
    # via
    #   -r requirements/req-prod.in
    #   djangorestframework
    #   model-bakery
djangorestframework==3.14.0
    # via -r requirements/req-prod.in
flake8==6.0.0
    # via -r requirements/req-dev.in
iniconfig==2.0.0
    # via pytest
mccabe==0.7.0
    # via flake8
model-bakery==1.12.0
    # via -r requirements/req-dev.in
mypy-extensions==1.0.0
    # via black
packaging==23.1
    # via
    #   black
    #   build
    #   pytest
pathspec==0.11.1
    # via black
pip-tools==7.0.0
    # via -r requirements/req-dev.in
platformdirs==3.9.1
    # via black
pluggy==1.2.0
    # via pytest
psycopg2-binary==2.9.6
    # via -r requirements/req-prod.in
pycodestyle==2.10.0
    # via flake8
pydantic==2.0.3
    # via -r requirements/req-prod.in
pydantic-core==2.3.0
    # via pydantic
pyflakes==3.0.1
    # via flake8
pyproject-hooks==1.0.0
    # via build
pytest==7.4.0
    # via
    #   -r requirements/req-dev.in
    #   pytest-django
    #   pytest-mock
    #   pytest-unordered
pytest-django==4.5.2
    # via -r requirements/req-dev.in
pytest-mock==3.11.1
    # via -r requirements/req-dev.in
pytest-unordered==0.5.2
    # via -r requirements/req-dev.in
pytz==2023.3
    # via djangorestframework
sqlparse==0.4.4
    # via django
typing-extensions==4.7.1
    # via
    #   pydantic
    #   pydantic-core
wheel==0.40.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
